packages/python/pyproject.toml
[project]
name = "dendrux"
version = "0.1.0a1"
description = "The runtime for agents that act in the real world."
readme = "README.md"
license = "Apache-2.0"
requires-python = ">=3.11"
authors = [
    { name = "Dendrux Contributors" },
]
keywords = ["agents", "llm", "ai", "tools", "runtime"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed",
]

dependencies = [
    "pydantic>=2.0,<3.0",
    "typer>=0.12,<1.0",
    "rich>=13.0,<15.0",
    "python-ulid>=2.0,<4.0",
]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40,<1.0"]
openai = ["openai>=1.50,<3.0"]
bridge = ["fastapi>=0.135,<1.0", "uvicorn[standard]"]
db = ["sqlalchemy[asyncio]>=2.0,<3.0", "aiosqlite>=0.19,<1.0", "alembic>=1.13,<2.0"]
postgres = ["asyncpg>=0.29,<1.0"]
all = ["dendrux[anthropic,openai,bridge,db,postgres]"]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.8",
    "mypy>=1.10",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.scripts]
dendrux = "dendrux.cli.main:app"

[project.urls]
Homepage = "https://github.com/dendrux/dendrux"
Repository = "https://github.com/dendrux/dendrux"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/dendrux"]

[tool.hatch.build.targets.wheel.force-include]
"src/dendrux/dashboard/static" = "dendrux/dashboard/static"

# --- Tool configs ---

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["src/dendrux/db/migrations"]

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "F",    # pyflakes
    "I",    # isort (import sorting)
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "SIM",  # flake8-simplify
    "TCH",  # flake8-type-checking
]

[tool.ruff.lint.isort]
known-first-party = ["dendrux"]

[tool.mypy]
strict = true
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[[tool.mypy.overrides]]
module = "anthropic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "openai.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "sqlalchemy.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aiosqlite.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "alembic.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "dendrux.db.migrations.*"
ignore_errors = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.coverage.run]
omit = [
    "src/dendrux/cli/*",
    "src/dendrux/db/migrations/*",
]
